refactor: ScrollableBlurArea styles/@common으로 분리
