--- frontend/src/components/@common/scrollableBlurArea/ScrollableBlurArea.tsx
@@ -1,11 +1,0 @@
-import * as S from './ScrollableBlurArea.styles';
-
-interface ScrollableBlurAreaProps {
-  isAtPageBottom: boolean;
-}
-
-const ScrollableBlurArea = ({ isAtPageBottom }: ScrollableBlurAreaProps) => {
-  return <S.BottomBlurArea $isHide={isAtPageBottom} />;
-};
-
-export default ScrollableBlurArea;
